refactor(config): 更新定时任务配置并优化进程终止逻辑
- 启用 curl 和 echo 定时任务，禁用 date 定时任务
- 修改 curl 任务的执行周期为每 10 秒一次
- 优化 terminate_rcron_process 函数：
  - 将 pid 类型从 i32 改为 u32
  - 移除不必要的类型转换

+++ src/main.rs
@@ -112,10 +112,10 @@ async fn terminate_rcron_process() -> Result<(), std::io::Error> {
         .await?;
 
     let pid_list = String::from_utf8_lossy(&output.stdout);
-    let pids = pid_list.split('\n').filter_map(|pid| pid.parse::<i32>().ok());
+    let pids = pid_list.split('\n').filter_map(|pid| pid.parse::<u32>().ok());
 
     for pid in pids {
-        if pid == current_pid as i32 {
+        if pid == current_pid {
             continue;
         }
         Command::new("kill")
